+++ jest/setup.js
@@ -1,7 +1,5 @@
 import 'react-native-gesture-handler/jestSetup';
 
-global.__reanimatedWorkletInit = jest.fn();
-
 jest.mock('react-native-reanimated', () => {
   const Reanimated = require('react-native-reanimated/mock');
 
@@ -16,3 +14,24 @@ global.__reanimatedWorkletInit = jest.fn();
 
 jest.mock('react-native/Libraries/Animated/NativeAnimatedHelper');
 jest.mock('react-native-mmkv');
+
+const mockedNavigation = jest.fn();
+
+jest.mock('@react-navigation/native', () => {
+  return {
+    useNavigation: () => ({
+      navigate: mockedNavigation,
+    }),
+  };
+});
+
+jest.mock('react-native-qrcode-scanner/node_modules/react-native-permissions', () =>
+  require('react-native-permissions/mock'),
+);
+
+jest.mock('react-native/Libraries/Components/Switch/Switch', () => {
+  const mockComponent = require('react-native/jest/mockComponent');
+  return {
+    default: mockComponent('react-native/Libraries/Components/Switch/Switch'),
+  };
+});
