+++ src/ui/screens/SendFundScreen.test.tsx
@@ -17,49 +17,54 @@ const route = {
   },
 } as RouteProp<AccountsStackParamList, typeof sendFundScreen>;
 
-describe('SendFundScreen component', () => {
-  it('On component loaded successfully', async () => {
-    const {getAllByText, queryByText, getByText, getByA11yState} = await waitFor(() =>
-      render(<SendFundScreen navigation={navigation} route={route} />),
-    );
-    expect(getAllByText('Send funds').length).toBe(1);
-    getAllByText('Enter amount');
-    getAllByText('Send to address');
-    getAllByText('Existential deposit');
-    getAllByText('Transfer with account keep-alive checks');
-    expect(queryByText('Normal transfer without keep-alive checks')).toBe(null);
-    expect(getByText(/Make Transfer/i)).toBeDisabled();
-    getByA11yState({disabled: true});
-  });
+test('Component loaded successfully with the "Make transfer" button disabled', async () => {
+  const {getAllByText, queryByText, getByText, getByA11yState} = await waitFor(() =>
+    render(<SendFundScreen navigation={navigation} route={route} />),
+  );
+  const navigationSpy = jest.spyOn(navigation, 'goBack');
+  expect(getAllByText('Send funds').length).toBe(1);
+  getAllByText('Enter amount');
+  getAllByText('Send to address');
+  getAllByText('Existential deposit');
+  getAllByText('Transfer with account keep-alive checks');
+  expect(queryByText('Normal transfer without keep-alive checks')).toBe(null);
+  expect(getByText(/Make Transfer/i)).toBeDisabled();
+  getByA11yState({disabled: true});
 
-  it('When user enter correct values and checked keep alive check', async () => {
-    const {queryAllByText, getByPlaceholderText, getByText} = await waitFor(() =>
-      render(<SendFundScreen navigation={navigation} route={route} />),
-    );
-    fireEvent.changeText(getByPlaceholderText('Enter amount'), '2588');
-    fireEvent.changeText(getByPlaceholderText('Account address'), 'test');
-    expect(queryAllByText('Enter a valid address').length).toEqual(1);
-    fireEvent.changeText(getByPlaceholderText('Account address'), '14yx4vPAACZRhoDQm1dyvXD3QdRQyCRRCe5tj1zPomhhS29a');
-    expect(queryAllByText('Enter a valid address').length).toEqual(0);
-    getByText('Transfer with account keep-alive checks');
-    const button = getByText('Make Transfer');
-    fireEvent.press(button);
-  });
+  fireEvent.press(getByText('Cancel'));
+  expect(navigationSpy).toBeCalledTimes(1);
+});
 
-  it('When user enter correct values and normal transfer', async () => {
-    const {queryAllByText, getByPlaceholderText, getByText, getByTestId, queryByText} = await waitFor(() =>
-      render(<SendFundScreen navigation={navigation} route={route} />),
-    );
-    fireEvent.changeText(getByPlaceholderText('Enter amount'), '2588');
-    fireEvent.changeText(getByPlaceholderText('Account address'), 'test');
-    expect(queryAllByText('Enter a valid address').length).toEqual(1);
-    fireEvent.changeText(getByPlaceholderText('Account address'), '14yx4vPAACZRhoDQm1dyvXD3QdRQyCRRCe5tj1zPomhhS29a');
-    expect(queryAllByText('Enter a valid address').length).toEqual(0);
-    const switchComponent = getByTestId('keep_alive_switch');
-    fireEvent(switchComponent, 'valueChange', false);
-    await waitFor(() => queryByText('Normal transfer without keep-alive checks'));
-    expect(queryByText('Transfer with account keep-alive checks')).toBe(null);
-    const button = getByText('Make Transfer');
-    fireEvent.press(button);
-  });
+test('When user enter correct values and checked keep alive check to make transfer', async () => {
+  const {queryAllByText, getByPlaceholderText, getByText} = await waitFor(() =>
+    render(<SendFundScreen navigation={navigation} route={route} />),
+  );
+
+  fireEvent.changeText(getByPlaceholderText('Enter amount'), '2588');
+
+  fireEvent.changeText(getByPlaceholderText('Account address'), 'test');
+  expect(queryAllByText('Enter a valid address').length).toEqual(1);
+  fireEvent.changeText(getByPlaceholderText('Account address'), '14yx4vPAACZRhoDQm1dyvXD3QdRQyCRRCe5tj1zPomhhS29a');
+  expect(queryAllByText('Enter a valid address').length).toEqual(0);
+
+  getByText('Transfer with account keep-alive checks');
+  const button = getByText('Make Transfer');
+  fireEvent.press(button);
+});
+
+test('When user enter correct values and normal transfer to make transfer', async () => {
+  const {queryAllByText, getByPlaceholderText, getByText, getByTestId, queryByText} = await waitFor(() =>
+    render(<SendFundScreen navigation={navigation} route={route} />),
+  );
+  fireEvent.changeText(getByPlaceholderText('Enter amount'), '2588');
+  fireEvent.changeText(getByPlaceholderText('Account address'), 'test');
+  expect(queryAllByText('Enter a valid address').length).toEqual(1);
+  fireEvent.changeText(getByPlaceholderText('Account address'), '14yx4vPAACZRhoDQm1dyvXD3QdRQyCRRCe5tj1zPomhhS29a');
+  expect(queryAllByText('Enter a valid address').length).toEqual(0);
+  const switchComponent = getByTestId('keep_alive_switch');
+  fireEvent(switchComponent, 'valueChange', false);
+  await waitFor(() => queryByText('Normal transfer without keep-alive checks'));
+  expect(queryByText('Transfer with account keep-alive checks')).toBe(null);
+  const button = getByText('Make Transfer');
+  fireEvent.press(button);
 });
